Feature/0012 — Narrow unit-quantity helpers to catalogue Price only
Confirmed by the existing OrderLineTest ($orderLine->decimal('unit_price')
returns 5.07 for unit_price=507, unit_quantity=100): `OrderLine.unit_price`
is persisted *already divided* to per-single-unit by CalculateLineSubtotal,
and `unit_quantity` is metadata for accounting — not a divisor for display.
v1 `Casts\Price` resolved `unit_quantity` for OrderLine the same way, which
meant `$line->unit_price->unitDecimal()` quietly double-divided whenever
`unit_quantity > 1`. The previous commit on this branch faithfully restored
that bug; this one removes it.

- Drop `unitDecimal()`/`unitFormat()` from `FormatsPrices`. The trait now
  exposes only `decimal()`/`format()` — which is the correct primitive for
  every model the trait targets.
- Add `unitDecimal()`/`unitFormat()` directly to the catalogue
  `Lunar\Core\Models\Price` model, where the unit-quantity division is
  genuinely needed (price stored as "£X for N units", `unit_quantity` lives
  on `priceable`).
- Narrow `RewriteUnitPriceFormatterCallRector` to a new
  `LunarSetList::V1_TO_V2_UNIT_AWARE_ATTRIBUTES` map (catalogue Price only).
  Unit calls on Order/OrderLine/Transaction are deliberately left untouched
  for manual review — the docblock and upgrade docs flag this.
- Fix `OrderItemsTable` (admin Filament) which constructed
  `PriceFormatterInterface` with `unitQty => unit_quantity` — the same
  double-divide on the read side. Now uses `$record->format('unit_price',
  decimalPlaces: 4)`.
- Add `compare_price` to the catalogue Price entry in
  `V1_TO_V2_MONEY_ATTRIBUTES` so the plain `format`/`decimal` rules cover
  it too.
- Replace OrderLine unit fixtures with catalogue-Price ones plus a
  negative fixture showing OrderLine unit calls are left for manual
  review.
