+++ tests/upgrade/Unit/Rector/Price/Fixture/RewriteUnitPriceFormatterCallRector/catalogue_price_unit_decimal.php.inc
@@ -1,0 +1,21 @@
+<?php
+
+namespace Lunar\Tests\Upgrade\Unit\Rector\Price\Fixture\RewriteUnitPriceFormatterCallRector;
+
+function unit_decimal(\Lunar\Core\Models\Price $catalogPrice): ?float
+{
+    return $catalogPrice->price->unitDecimal();
+}
+
+?>
+-----
+<?php
+
+namespace Lunar\Tests\Upgrade\Unit\Rector\Price\Fixture\RewriteUnitPriceFormatterCallRector;
+
+function unit_decimal(\Lunar\Core\Models\Price $catalogPrice): ?float
+{
+    return $catalogPrice->unitDecimal('price');
+}
+
+?>
