+++ tests/upgrade/Unit/Rector/Price/config/unit.php
@@ -7,4 +7,4 @@ use Lunar\Upgrade\Rector\Price\RewriteUnitPriceFormatterCallRector;
 use Rector\Config\RectorConfig;
 
 return RectorConfig::configure()
-    ->withConfiguredRule(RewriteUnitPriceFormatterCallRector::class, LunarSetList::V1_TO_V2_MONEY_ATTRIBUTES);
+    ->withConfiguredRule(RewriteUnitPriceFormatterCallRector::class, LunarSetList::V1_TO_V2_UNIT_AWARE_ATTRIBUTES);
